Renamed remaining packages
