+++ rmf_visualization/package.xml
@@ -9,7 +9,7 @@
 
   <buildtool_depend>ament_cmake</buildtool_depend>
 
-  <exec_depend>rmf_schedule_visualizer</exec_depend>
+  <exec_depend>rmf_visualization_schedule</exec_depend>
   <exec_depend>rmf_visualization_fleet_states</exec_depend>
   <exec_depend>rmf_visualization_fleet_building_systems</exec_depend>
   <exec_depend>rmf_visualization_rviz2_plugins</exec_depend>
