+++ rmf_visualization_rviz2_plugins/package.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>rviz2_plugin</name>
+  <name>rmf_visualization_rviz2_plugins</name>
   <version>1.1.0</version>
-  <description>A package to create a custom plugin for rviz</description>
+  <description>A package to create a custom plugins for rviz2</description>
   <maintainer email="yadunund@openrobotics.org">yadu</maintainer>
   <license>Apache License 2.0</license>
 
@@ -12,7 +12,7 @@
   <depend>rclcpp</depend>
   <depend>rmf_door_msgs</depend>
   <depend>rmf_lift_msgs</depend>
-  <depend>rmf_schedule_visualizer_msgs</depend>
+  <depend>rmf_visualization_msgs</depend>
   <depend>rmf_traffic_ros2</depend>
   <depend>rviz_common</depend>
   <depend>rviz_rendering</depend>
