+++ rmf_visualization_rviz2_plugins/plugin_description.xml
@@ -1,20 +1,20 @@
-<library path="rviz2_plugin">
-  <class name="rviz2_plugin/SchedulePanel"
-         type="rviz2_plugin::SchedulePanel"
+<library path="rmf_visualization_rviz2_plugins">
+  <class name="rmf_visualization_rviz2_plugins/SchedulePanel"
+         type="rmf_visualization_rviz2_plugins::SchedulePanel"
          base_class_type="rviz_common::Panel">
     <description>
       A panel widget allowing users to adjust parameters in the schedule visualizaer.
     </description>
   </class>
-  <class name="rviz2_plugin/LiftPanel"
-      type="rviz2_plugin::LiftPanel"
+  <class name="rmf_visualization_rviz2_plugins/LiftPanel"
+      type="rmf_visualization_rviz2_plugins::LiftPanel"
       base_class_type="rviz_common::Panel">
     <description>
       A panel widget allowing users to send lift requests and visualize states.
     </description>
   </class>
-  <class name="rviz2_plugin/DoorPanel"
-      type="rviz2_plugin::DoorPanel"
+  <class name="rmf_visualization_rviz2_plugins/DoorPanel"
+      type="rmf_visualization_rviz2_plugins::DoorPanel"
       base_class_type="rviz_common::Panel">
     <description>
       A panel widget allowing users to send door requests and visualize states.
