+++ rmf_visualization_rviz2_plugins/src/DoorPanel.cpp
@@ -22,7 +22,7 @@
 #include "DoorPanel.hpp"
 #include "StandardNames.hpp"
 
-namespace rviz2_plugin {
+namespace rmf_visualization_rviz2_plugins {
 
 //==============================================================================
 
@@ -373,7 +373,7 @@ QString DoorPanel::door_state_mode_tooltip() const
 
 //==============================================================================
 
-} // namespace rviz2_plugin
+} // namespace rmf_visualization_rviz2_plugins
 
 #include <pluginlib/class_list_macros.hpp>
-PLUGINLIB_EXPORT_CLASS(rviz2_plugin::DoorPanel, rviz_common::Panel)
+PLUGINLIB_EXPORT_CLASS(rmf_visualization_rviz2_plugins::DoorPanel, rviz_common::Panel)
