+++ rmf_visualization_rviz2_plugins/src/SchedulePanel.cpp
@@ -27,7 +27,7 @@
 
 #include <rmf_traffic_ros2/StandardNames.hpp>
 
-namespace rviz2_plugin {
+namespace rmf_visualization_rviz2_plugins {
 
 using namespace std::chrono_literals;
 
@@ -335,7 +335,7 @@ void SchedulePanel::load(const rviz_common::Config& config)
   }
 }
 
-} // namespace rviz2_plugin
+} // namespace rmf_visualization_rviz2_plugins
 
 #include <pluginlib/class_list_macros.hpp>
-PLUGINLIB_EXPORT_CLASS(rviz2_plugin::SchedulePanel, rviz_common::Panel)
+PLUGINLIB_EXPORT_CLASS(rmf_visualization_rviz2_plugins::SchedulePanel, rviz_common::Panel)
